change tol=1e-9 and maxiter=100
the reason is that if tolerance is too small, it becomes strict for floating-point operations on GPUs, some values could become invalid without proper detection. A large maxiter combined with tight tolerance, the solver might be oscillating or diverging on some iteractions, later iteractions could be working with increasingly unstable values. This creates conditions where memory becomes corrupted.
The problem occrus in the norm function, related to this line: return CUDA.@allowscalar sqrt(r[1]). CUDA attempts to copy a scalar value from GPU to CPU. If the value is corrupted due to numerical instabiltiy, the copy fails and results in the illegal memory access error.
Therefore, setting reltol = 1e-9 and maxiter=100 should work.